Service Impl Updated

+++ Home Loan App Sprint-1 Git/src/main/java/com/ja5g4/homeloan/repository/ILoanAgreementRepository.java
@@ -3,7 +3,7 @@ package com.ja5g4.homeloan.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.ja5g4.homeloan.entities.LoanAgreement;
 
-public interface ILoanAgreementRepository extends JpaRepository<LoanAgreement, Integer> {
+public interface ILoanAgreementRepository extends JpaRepository<LoanAgreement, Long> {
 	
 
 }
