+++ Home Loan App Sprint-1 Git/src/main/java/com/ja5g4/homeloan/repository/ILoanApplicationRepository.java
@@ -3,7 +3,7 @@ package com.ja5g4.homeloan.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.ja5g4.homeloan.entities.LoanApplication;
 
-public interface ILoanApplicationRepository extends JpaRepository<LoanApplication, Integer> {
+public interface ILoanApplicationRepository extends JpaRepository<LoanApplication, Long> {
